proxy/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.27.2
openai==1.54.4
supabase==2.9.1
redis==5.2.0
numpy==2.1.3
python-dotenv==1.0.1
